[IMP] Improve security for the asynchronous export by linking the attachment to a record only visible to the user
