+++ base_export_async/data/cron.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
     <record id="to_delete_attachment" model="ir.cron">
-        <field name="name">Delete export generated attachment</field>
-        <field name="model_id" ref="model_ir_attachment"/>
+        <field name="name">Delete Generated Exports</field>
+        <field name="model_id" ref="model_delay_export"/>
         <field name="state">code</field>
         <field name="code">model.cron_delete()</field>
         <field name='interval_number'>1</field>
