+++ base_export_async/models/__init__.py
@@ -1,5 +1,4 @@
 # Copyright 2019 ACSONE SA/NV
 # License AGPL-3.0 or later (http://www.gnu.org/licenses/agpl).
 
-from . import attachment
 from . import delay_export
